LateNight Changes
Just Pushed a little changed, made it so that when you leave the jail cell you go to the hallway, aswell as when you are in the courtyard and you grab the key and leave the gate you will be put back in the hallway, also the PINKISH door is to the courtyard and the GRAY door goes to the new security room not fin yet. Also added two famalies which is so far all the player sprites and the MainCell HitBoxes, So care when chaging anything around them just rember there in a family when coding a change.