12/16 | Bug Fixes | Last bug is the door spawns
Saving and pushing now guys, Got the Press E to come up and also should be able to close the text when ever you want and come back to it if you miss click an hit E to fast. I changed some more stuff in the Kitchen like some pans and one has the key hidden under so that ppl who played our game in the aplha dont just remeber and skip to the key instead of reading the guys text also i will leave the keys loactaion hidden as a suprise for you guys to. In the common room there are now two routes to take to get the the key top side and bottom side. Overall The real only bug left would be the door spawn but thats not a huge issue at all because the entire game works so far lol atleast for the most part. I tried it tonight ran pretty good all the way to 5 keys and end game. Gn see yall 2moro